Revert "Add Project Download"
