+++ client/app/project/toolbar/toolbar.html
@@ -12,8 +12,9 @@
         <ul class='dropdown-menu'>
           <li ng-click="openAddFileModal()">Add File</li>
           <li ng-click="openAddFolderModal()">Add Folder</li>
-          <li ng-click='downloadFile()'>Download Current File</li>
-          <li ng-click='downloadProjectZip()'>Download Project</li>
+          <li ng-click='downloadFile()'>
+            <a>Download Current File</a>
+          </li>
         </ul>
       </li>
       <li class='dropdown' dropdown>
