+++ package.json
@@ -6,8 +6,7 @@
   "scripts": {
     "start": "./node_modules/forever/bin/forever start -p ./data/logs -a -o ./data/logs/stdout.log -e ./data/logs/stderr.log server ",
     "test": "export NODE_ENV='test'; ./node_modules/mocha/bin/mocha -t 5000 server/tests/index",
-    "dev": "./node_modules/nodemon/bin/nodemon.js server",
-    "clean": "npm install && bower install && node server/deleteAllDatabases && gulp"
+    "dev": "./node_modules/nodemon/bin/nodemon.js server"
   },
   "repository": {
     "type": "git",
