+++ server/file/fileController.js
@@ -14,6 +14,7 @@ var mongoIndex = function (str) {
 
 var fileController = {
   createNewFileOrFolder: function (req, res) {
+    console.log('createNewFileOrFolder');
     var projectName = req.body.project_name || req.param('project_name');
     var fileName = req.body.file_name || req.param('file_name');
     var type = req.body.type || req.param('type');
