+++ server/project/projectController.js
@@ -2,12 +2,13 @@
 var Promise = require('bluebird');
 var db = require('../db');
 var _ = require('lodash');
+var Q = require('q');
 
 var models = require('../models.js').models;
+var getFileContents = require('../download/downloadController')._getFileContents;
 var getFileStructure = require('../file/fileController').getFileStructure;
-var getProject = require('./getProject');
-var getUser = require('./getUser');
-var getProjectZip = require('./getProjectZip');
+var getPathsForFileStructure = require('../file/fileController').getPathsForFileStructure;
+var JSZip = require("jszip");
 
 var projectController = {};
 
@@ -67,7 +68,18 @@ projectController.getAllProjects = function (req, res) {
  */
 projectController.getSpecificProject = function (req, res) {
   //only get the requested project if the user has a relation with it
-  return getProject(req.params.project_name_or_id)
+  models.Project
+    .query({
+      where: {
+        project_name: req.params.project_name_or_id
+      },
+      orWhere: {
+        id: req.params.project_name_or_id
+      }
+    })
+    .fetch({
+      withRelated: ['user']
+    })
     .then(function (project) {
       return getFileStructure(project.get('id'))
         .then(function (fileStructure) {
@@ -91,7 +103,15 @@ projectController.addUser = function (req, res) {
   var project_name = req.body.project_name;
   var newUserName = req.body.newUserName;
   //only only add the user if the person that requested the addition is a listed user of the project
-  return getUser(newUserName)
+  models.User
+    .query({
+      where: {
+        username: newUserName
+      }
+    })
+    .fetch({
+      withRelated: ['project']
+    })
     .then(function (user) {
       if (!user) throw new Error('There is not model with this name');
       // Change so that if there is not user, it does a res.end saying 'there is no user'
@@ -99,7 +119,15 @@ projectController.addUser = function (req, res) {
       return user;
     })
     .then(function (queriedUser) {
-      return getProject(project_name)
+      return models.Project
+        .query({
+          where: {
+            project_name: project_name
+          }
+        })
+        .fetch({
+          withRelated: ['user']
+        })
         .then(function (model) {
           return model
             .related('user')
@@ -128,7 +156,13 @@ projectController.addUser = function (req, res) {
  */
 projectController.delete = function (req, res) {
   //only delete the project if person requesting the deletion has a relation with it
-  return getProject(req.body.id)
+  models.Project
+    .query({
+      where: {
+        id: req.body.id
+      }
+    })
+    .fetch()
     .then(function (model) {
       if (model === null) {
         throw new Error('No model found for that id');
@@ -155,14 +189,53 @@ projectController.delete = function (req, res) {
  * Download a project as a .zip
  */
 projectController.downloadSpecificProject = function (req, res) {
-  return getProjectZip(req.params.project_name_or_id)
-    .then(function (zipObject) {
-      res.setHeader('Content-disposition', 'attachment; filename=' + zipObject.name);
+  var project;
+   return models.Project
+    .query({
+      where: {
+        project_name: req.params.project_name_or_id
+      },
+      orWhere: {
+        id: req.params.project_name_or_id
+      }
+    })
+    .fetch()
+    .then(function (_project) {
+      project = _project;
+      if (!project) throw new Error('No Model Could Be Found');
+      return getFileStructure(project.get('id'), project.get('project_name'))
+         .then(function (fileStructure) {
+          var paths = getPathsForFileStructure(fileStructure);
+          return Q.allSettled(paths.map(function (path) {
+            return getFileContents(project, path)
+              .then(function (fileContents) {
+                return {
+                  path: path,
+                  fileContents: fileContents
+                };
+              });
+          }));
+        });
+    })
+    .then(function (allFileContents){
+      var projectArchive = new JSZip();
+      allFileContents.forEach(function (file) {
+        var filePath = file.value.path;
+        var fileContents = file.value.fileContents;
+        projectArchive.file(filePath, fileContents);
+      });
+      var content = null;
+      if (JSZip.support.uint8array) {
+        content = projectArchive.generate({type : 'uint8array'});
+      } else {
+        content = projectArchive.generate({type : 'string'});
+      }
+      res.setHeader('Content-disposition', 'attachment; filename=' + project.get('project_name'));
       res.setHeader('content-type', 'application/zip');
-      res.sendFile(zipObject.path);
+      res.send(content);
     })
     .catch(function (err) {
-      console.log('Error Getting Project Zip', err);
+      console.log('ERROR', err);
       res.status(400).send(err);
     });
 };
