+++ server/tests/index.js
@@ -1,8 +1,11 @@
 /*global describe:true, before: true, after: true */
 'use strict';
 
+var config = require('config');
+var Promise = require('bluebird');
+var should = require('should');
 var db = require('../db');
-var deleteAllDatabases = require('../deleteAllDatabases');
+var mongoClient =  Promise.promisifyAll(require('mongodb').MongoClient);
 
 describe('Code Friends', function () {
 
@@ -18,8 +21,33 @@ describe('Code Friends', function () {
 
     // Delete All Test Tables
   after(function (done) {
-    deleteAllDatabases()
+    db.schema.dropTable('projects_users')
       .then(function () {
+        return Promise.all([
+          db.schema.dropTable('users'),
+          db.schema.dropTable('projects'),
+        ]);
+      })
+      .then(function () {
+        console.log('Deleting All MySQL Tables');
+        return true;
+      })
+      .then(function () {
+        return mongoClient.connectAsync(config.get('mongo'))
+          .then(function (db) {
+              var projectCollection = Promise.promisifyAll(db.collection('project_file_structre'));
+              return projectCollection.removeAsync();
+          })
+          .then(function () {
+            console.log('Deleting All Mongo Collections');
+            return true;
+          })
+          .catch(function (err) {
+            console.log('Error Deleting Mongo Collection', err);
+          });
+      })
+      .then(function () {
+        console.log('Deleting Database Done');
         done();
       })
       .catch(function (err) {
